refactor(web): 모임 생성 도메인 모델 entities/moim으로 분리
